Workaround to make LCD work after reset.
The workaround is the delayms(1000) in main().

Since we introduced SLEEP mode a reset would crash the LCD. It seems the communication between the PIC and the LCD is disrupted after the reset and it never recovers. By waiting this 1s before the program fully restarts again the communication works well again.

A software reset after the SLEEP instruction worked fine. It seems the problem is when the reset occurs during SLEEP.

There might be a way to reset the LCD state. If so it should be tried to reset the LCD instead of using the current workaround.